Element and attribute name changes, additional element
Added the "question" element to incorporate peer engagement; as student writers are using this schema to engage with each others' text, the "question" element will encourage them to think about questions they might have. Also added an attribute called "playType" (in <general> tag under <original>) with the different play elements as the values so that the writer may specify which type of play element to which they are connecting this moment.